done some exercises

+++ exercises/060_floats.zig
@@ -43,7 +43,7 @@ pub fn main() void {
     //
     // We'll convert this weight from tons to kilograms at a
     // conversion of 907.18kg to the ton.
-    const shuttle_weight: f16 = 907.18 * 2200;
+    const shuttle_weight: f32 = 907.18 * 2200;
 
     // By default, float values are formatted in scientific
     // notation. Try experimenting with '{d}' and '{d:.3}' to see
